Fix handling of fractional numbers
According to protocol guide the timestamp should always be expressed in
milliseconds. Manyverse also follows this standard.

This type contained some kind of a fix where in some cases the number was
treated differently. The new version always treats the number as milliseconds,
I just discarded the part after the dot as we don't need such precision.